REF np.ndarray[type, ndim, ...] -> np.ndarray
We no longer need to use the buffer syntax, as all arrays are passed
through PyArray_FROMANY which guarantess that the array has the right
properties. This works around NumPy setting F_CONTIGUOUS to False even
if an array is only reshaped with 1-length dims (needed in a commit
soon to follow).